Renamed  a few packages in the toon binding. See openhab/static-code-analysis#100
