+++ addons/binding/org.openhab.binding.toon/src/main/java/org/openhab/binding/toon/handler/ToonBridgeHandler.java
@@ -24,9 +24,9 @@ import org.eclipse.smarthome.core.thing.binding.BaseBridgeHandler;
 import org.eclipse.smarthome.core.thing.binding.ThingHandler;
 import org.eclipse.smarthome.core.types.Command;
 import org.eclipse.smarthome.core.types.RefreshType;
-import org.openhab.binding.toon.config.ToonBridgeConfiguration;
 import org.openhab.binding.toon.internal.ToonApiClient;
 import org.openhab.binding.toon.internal.api.ToonState;
+import org.openhab.binding.toon.internal.config.ToonBridgeConfiguration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
