+++ addons/binding/org.openhab.binding.toon/src/main/java/org/openhab/binding/toon/internal/ToonHandlerFactory.java
@@ -18,10 +18,10 @@ import org.eclipse.smarthome.core.thing.Thing;
 import org.eclipse.smarthome.core.thing.ThingTypeUID;
 import org.eclipse.smarthome.core.thing.binding.BaseThingHandlerFactory;
 import org.eclipse.smarthome.core.thing.binding.ThingHandler;
-import org.openhab.binding.toon.discovery.ToonDiscoveryService;
 import org.openhab.binding.toon.handler.ToonBridgeHandler;
 import org.openhab.binding.toon.handler.ToonDisplayHandler;
 import org.openhab.binding.toon.handler.ToonPlugHandler;
+import org.openhab.binding.toon.internal.discovery.ToonDiscoveryService;
 import org.osgi.framework.ServiceRegistration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
