+++ addons/binding/org.openhab.binding.toon/src/main/java/org/openhab/binding/toon/internal/discovery/ToonDiscoveryService.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.toon.discovery;
+package org.openhab.binding.toon.internal.discovery;
 
 import static org.openhab.binding.toon.ToonBindingConstants.*;
